first go on trying to make this NN learnable. Still get weird outputs

+++ Src/TrainingData.java
@@ -1,10 +1,10 @@
 package Src;
 
 public class TrainingData {
-    double inputData[]; // x Vector
-    double expectedResult[]; // y Vector
+    float inputData[]; // array because the input data will be a Vector (x input leads to y output)
+    float expectedResult[]; // array because we have a Vector of expected results
 
-    public TrainingData(double inputData[], double expectedResult[]) {
+    public TrainingData(float inputData[], float expectedResult[]) {
         this.inputData = inputData;
         this.expectedResult = expectedResult;
     }
